+++ src/App.js
@@ -15,6 +15,7 @@ import ScrollToTop from 'components/ScrollToTop';
 import { Alert } from 'reactstrap';
 import env from '@beam-australia/react-env';
 import { Helmet } from 'react-helmet';
+import ErrorBoundary from 'components/ErrorBoundary/ErrorBoundary';
 
 const alertStyle = { borderRadius: '0', marginTop: '-30px', marginBottom: '30px' };
 
@@ -25,27 +26,29 @@ const App = () => {
 
     return (
         <ScrollToTop>
-            <DefaultLayout>
-                {showBrowserWarning && (
-                    <Alert color="danger" style={alertStyle} className="text-center">
-                        <strong>Outdated browser</strong> You are using Internet Explorer which is not supported. Please upgrade your browser for the
-                        best experience
-                    </Alert>
-                )}
-                {env('IS_TESTING_SERVER') === 'true' && (
-                    <>
-                        <Helmet>
-                            <meta name="robots" content="noindex" /> {/* make sure search engines are not indexing our test server */}
-                        </Helmet>
-                        <Alert color="warning" style={alertStyle} className="text-center">
-                            <strong>Warning:</strong> You are using a testing environment. Data you enter in the system can be deleted without any
-                            notice.
+            <ErrorBoundary>
+                <DefaultLayout>
+                    {showBrowserWarning && (
+                        <Alert color="danger" style={alertStyle} className="text-center">
+                            <strong>Outdated browser</strong> You are using Internet Explorer which is not supported. Please upgrade your browser for
+                            the best experience
                         </Alert>
-                    </>
-                )}
-                {/* Suspense is used for when the component is lazy loaded */}
-                <Suspense fallback={<div className="mt-5 mb-2 text-center">Loading...</div>}>{element}</Suspense>
-            </DefaultLayout>
+                    )}
+                    {env('IS_TESTING_SERVER') === 'true' && (
+                        <>
+                            <Helmet>
+                                <meta name="robots" content="noindex" /> {/* make sure search engines are not indexing our test server */}
+                            </Helmet>
+                            <Alert color="warning" style={alertStyle} className="text-center">
+                                <strong>Warning:</strong> You are using a testing environment. Data you enter in the system can be deleted without any
+                                notice.
+                            </Alert>
+                        </>
+                    )}
+                    {/* Suspense is used for when the component is lazy loaded */}
+                    <Suspense fallback={<div className="mt-5 mb-2 text-center">Loading...</div>}>{element}</Suspense>
+                </DefaultLayout>
+            </ErrorBoundary>
         </ScrollToTop>
     );
 };