+++ src/constants/graphSettings.js
@@ -95,7 +95,7 @@ export const CLASSES = {
     QB_DATASET_CLASS: 'QBDataset',
     LOCATION: 'DCLocation',
     SENTENCE: 'Sentence',
-    BENCHMARK: 'Benchmark',
+    BENCHMARK: 'C14022',
     MODEL: 'Model',
     DATASET: 'Dataset',
     EVALUATION: 'Evaluation',
@@ -122,6 +122,7 @@ export const CLASSES = {
     INTEGER: 'Integer',
     BOOLEAN: 'Boolean',
     URI: 'URI',
+    PAPER_VERSION: 'PaperVersion',
     SOFTWARE: 'Software',
     EXTERNAL: 'External',
     DIAGRAM: 'Diagram',
@@ -130,6 +131,7 @@ export const CLASSES = {
 export const RESOURCES = {
     RESEARCH_FIELD_MAIN: 'R11',
     RESEARCH_FIELD_COMPUTER_SCIENCE: 'R132',
+    RESEARCH_FIELD_COMPUTATIONAL_LINGUISTICS: 'R322',
     EMPTY_RESOURCE: 'empty',
 };