+++ src/routes.config.js
@@ -11,7 +11,7 @@ import ComparisonDiff from 'pages/Comparisons/ComparisonDiff';
 import Home from 'pages/Home';
 import Changelog from 'pages/Changelog/Changelog';
 import NotFound from 'pages/NotFound';
-import Papers from 'pages/Papers';
+import Papers from 'pages/Papers/Papers';
 import Comparisons from 'pages/Comparisons/Comparisons';
 import Visualizations from 'pages/Visualizations/Visualizations';
 import Visualization from 'pages/Visualizations/Visualization';
@@ -35,11 +35,9 @@ import AddOrganization from 'pages/Organizations/AddOrganization';
 import AddObservatory from 'pages/Observatories/AddObservatory';
 import Observatory from 'pages/Observatories/Observatory';
 import SearchResults from 'pages/Search';
-import ViewPaper from 'pages/ViewPaper';
 import Stats from 'pages/Stats';
 import UserSettings from 'pages/UserSettings';
 import UserProfile from 'pages/UserProfile';
-import FeaturedComparisons from 'pages/FeaturedComparisons';
 import Data from 'pages/Data';
 import Contribution from 'pages/Contribution';
 import CsvImport from 'pages/CsvImport';
@@ -61,6 +59,7 @@ import HelpCenterCategory from 'pages/HelpCenter/HelpCenterCategory';
 import HelpCenterArticle from 'pages/HelpCenter/HelpCenterArticle';
 import HelpCenterSearch from 'pages/HelpCenter/HelpCenterSearch';
 import WebinarMay11 from 'pages/WebinarMay11';
+import CheckPaperVersion from 'pages/Papers/CheckPaperVersion';
 import Lists from 'pages/Lists/Lists';
 import ListNew from 'pages/Lists/ListNew';
 import List from 'pages/Lists/List';
@@ -76,6 +75,7 @@ import ResearchProblem from './pages/ResearchProblem';
 // run "npm run analyze" to ensure the listed dependencies are not loaded elsewhere and thus end up in the bundle
 const PdfTextAnnotation = lazy(() => import('pages/PdfTextAnnotation')); // for dependency "react-pdf-highlighter" ~1.16MB
 const PdfAnnotation = lazy(() => import('pages/PdfAnnotation')); // for dependency "handsontable" ~887.4KB
+const FeaturedComparisons = lazy(() => import('pages/FeaturedComparisons')); // for dependency @fontawesome/free-solid-svg-icons used to show icons
 
 const routes = [
     {
@@ -149,11 +149,11 @@ const routes = [
     {
         /* TODO: slug for the paper title */
         path: ROUTES.VIEW_PAPER_CONTRIBUTION,
-        element: ViewPaper,
+        element: CheckPaperVersion,
     },
     {
         path: ROUTES.VIEW_PAPER,
-        element: ViewPaper,
+        element: CheckPaperVersion,
     },
     {
         path: ROUTES.COMPARISON_DIFF,