+++ src/services/mocks/backend/resources.js
@@ -112,6 +112,24 @@ const resources = [
             }),
         );
     }),
+    rest.put(`${resourcesUrl}:id`, async (req, res, ctx) => {
+        const { label, classes } = await req.json();
+        const { id } = req.params;
+        return res(
+            ctx.json({
+                id,
+                label,
+                created_at: '2020-06-22T10:38:53.178764Z',
+                classes: classes ?? [],
+                shared: 0,
+                created_by: '00000000-0000-0000-0000-000000000000',
+                _class: 'resource',
+                observatory_id: '00000000-0000-0000-0000-000000000000',
+                extraction_method: 'UNKNOWN',
+                organization_id: '00000000-0000-0000-0000-000000000000',
+            }),
+        );
+    }),
 ];
 
 export default resources;