+++ src/slices/addPaperSlice.js
@@ -279,13 +279,7 @@ export const createContributionAction = ({ selectAfterCreation = false, fillStat
     );
 
     if (selectAfterCreation) {
-        dispatch(
-            selectResource({
-                increaseLevel: false,
-                resourceId: newResourceId,
-                label: newContributionLabel,
-            }),
-        );
+        dispatch(selectContribution(newContributionId));
     }
 
     // Dispatch loading template of classes
@@ -368,6 +362,7 @@ export const getResourceObject = (data, resourceId, newProperties) => {
                     }
                     return {
                         '@id': newResources.includes(value.resourceId) ? `_${value.resourceId}` : value.resourceId,
+                        '@type': value._class,
                     };
                 }),
             };